Build the database/sql integration only where it can be used
HexColor.Value returns driver.Value, and that single return type is the whole
of this package's dependency on database/sql/driver. It costs more than it
looks: database/sql/driver pulls uuid, which pulls crypto/rand, which pulls
crypto/fips140. On a js/wasm build that is 26 crypto packages and 169KB of a
3.2MB binary, and neither js/wasm nor TinyGo has a database/sql driver for
HexColor to be a Valuer for.

Scan and Value move to hexcolor_sql.go behind !tinygo && !js, with the tests
that cover them. Every other target builds and behaves exactly as before;
String, the JSON and YAML marshallers and the envconfig Decoder are unchanged
and still built everywhere.
