+++ hexcolor_sql_test.go
@@ -1,0 +1,34 @@
+//go:build !tinygo && !js
+
+package colorful
+
+import (
+	"reflect"
+	"testing"
+)
+
+// TestHexColorSQL covers the database/sql integration, and so carries the same
+// build constraint as the code it tests.
+func TestHexColorSQL(t *testing.T) {
+	for _, tc := range []struct {
+		hc HexColor
+		s  string
+	}{
+		{HexColor{R: 0, G: 0, B: 0}, "#000000"},
+		{HexColor{R: 1, G: 0, B: 0}, "#ff0000"},
+		{HexColor{R: 0, G: 1, B: 0}, "#00ff00"},
+		{HexColor{R: 0, G: 0, B: 1}, "#0000ff"},
+		{HexColor{R: 1, G: 1, B: 1}, "#ffffff"},
+	} {
+		var gotHC HexColor
+		if err := gotHC.Scan(tc.s); err != nil {
+			t.Errorf("_.Scan(%q) == %v, want <nil>", tc.s, err)
+		}
+		if !reflect.DeepEqual(gotHC, tc.hc) {
+			t.Errorf("_.Scan(%q) wrote %v, want %v", tc.s, gotHC, tc.hc)
+		}
+		if gotValue, err := tc.hc.Value(); err != nil || !reflect.DeepEqual(gotValue, tc.s) {
+			t.Errorf("%v.Value() == %v, %v, want %v, <nil>", tc.hc, gotValue, err, tc.s)
+		}
+	}
+}
